services/notification-service/requirements.txt
flask==3.0.0
twilio==9.0.0
pika==1.3.2
python-dotenv==1.0.1
